v2/requirements.txt
# Core dependencies
manga-ocr>=0.1.11
argostranslate>=1.9.0
mss>=9.0.0
PyYAML>=6.0
Pillow>=10.0.0

# macOS only - window capture
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"

# Windows only - window capture
pygetwindow>=0.0.9; sys_platform == "win32"
